refactor: 패키지명 수정 및 DAO 관련 에러 메세지 개선
