+++ src/main/java/dao/Connector.java
@@ -1,4 +1,4 @@
-package repository;
+package dao;
 
 import java.sql.Connection;
 import java.sql.SQLException;
