+++ src/test/java/dao/MySQLConnectorTest.java
@@ -1,4 +1,4 @@
-package repository.mock;
+package dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,10 +6,9 @@ import java.sql.Connection;
 import java.sql.SQLException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import repository.Connector;
 
-class TestConnectorTest {
-    private final Connector connector = new TestConnector();
+public final class MySQLConnectorTest {
+    private final Connector connector = new MySQLConnector();
 
     @Test
     @DisplayName("MySQL DB 연결을 확인한다.")
