+++ src/test/java/dao/mock/TestConnectorTest.java
@@ -1,14 +1,15 @@
-package repository;
+package dao.mock;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import dao.Connector;
 import java.sql.Connection;
 import java.sql.SQLException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public final class MySQLConnectorTest {
-    private final Connector connector = new MySQLConnector();
+class TestConnectorTest {
+    private final Connector connector = new TestConnector();
 
     @Test
     @DisplayName("MySQL DB 연결을 확인한다.")
